requirements.txt
langchain_classic==1.0.7
langchain_community==0.4.1
langchain_core==1.4.0
langchain_deepseek==1.0.1
langchain_huggingface==1.2.2
langchain_text_splitters==1.1.2
langgraph==1.2.0
modelscope==1.36.1
pydantic==2.13.4
pymilvus==2.6.12
pymilvus_model==0.3.2
python-dotenv==1.2.2
Requests==2.34.0
torch==2.11.0+cu130
tqdm==4.67.3
transformers==5.6.2
typing_extensions==4.15.0
